+++ components/DogCardContainer/index.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
 
-import Carousel from '@components/Carousel';
+import Carousel from '@components/Common/Carousel';
 
 import { SpaceAround } from '@styles/common';
 import TestImage from '@assets/TestImage.png';
