+++ components/index.ts
@@ -3,7 +3,7 @@ export { default as Header } from './Header';
 export { default as Overlay } from './Overlay';
 export { default as SideBar } from './SideBar';
 export { default as MainPageInform } from './MainPageInform';
-export { default as DogCardContainer } from './Carousel/DogCardContainer';
+export { default as DogCardContainer } from './DogCardContainer';
 
 // SearchBar
 export { default as SearchBar } from './SearchBar';
@@ -13,3 +13,4 @@ export { default as SearchBarForm } from './SearchBarForm';
 // Common
 export { default as Footer } from './Common/Footer';
 export { default as DropdownSelector } from './Common/DropdownSelector';
+export { default as Carousel } from './Common/Carousel';
